refactor(frontend): create FormDate component
