+++ frontend/assets/js/components/hotelbooking/Room.js
@@ -7,12 +7,19 @@ import Button from "react-bootstrap/Button";
 import Image from "react-bootstrap/Image";
 
 const Room = ({ room, booking }) => {
-  const { id, name, description, location, image, price, owner_id, status } =
-    room;
-  // console.log(room);
+  const {
+    id,
+    room_name,
+    description,
+    location,
+    image,
+    price,
+    owner_id,
+    status,
+  } = room;
 
   const triggerBooking = () => {
-    booking(owner_id, name, price);
+    booking(owner_id, room_name, price);
     console.log("Called booking in triggerBooking");
   };
 
@@ -20,11 +27,11 @@ const Room = ({ room, booking }) => {
     <Row style={{ padding: "20px" }}>
       <Col xs={1}></Col>
       <Col xs={2}>
-        <Image src={image} alt={name} width='300' fluid />
+        <Image src={image} alt={room_name} width='300' fluid />
       </Col>
       <Col xs={7}>
         <h3>HOTEL NAME</h3>
-        <h4>{name}</h4>
+        <h4>{room_name}</h4>
         <p>{description}</p>
         <h5>{location}</h5>
         <p>1 room / 1 bed / 1 night</p>
